chore: Refactor screenshot script to extract parser into a file
