+++ scripts/screenshots.js
@@ -1,17 +1,11 @@
 #!/usr/bin/env node
 
-const { ArgumentParser } = require('argparse')
 const { prepareFS, prepareBrowser } = require('./screenshots/prepares')
 const screenshotReactStyleguide = require('./screenshots/screenshotReactStyleguide')
 const screenshotStackExamples = require('./screenshots/screenshotStackExamples')
 const screenshotKSSStyleguide = require('./screenshots/screenshotKSSStyleguide')
-const {
-  readConfig,
-  builtinViewports,
-  parseViewportArgument,
-  pathArgument,
-  urlArgument
-} = require('./screenshots/helpers')
+const makeParser = require('./screenshots/parser')
+const { readConfig, parseViewportArgument } = require('./screenshots/helpers')
 
 let puppeteer
 try {
@@ -28,45 +22,7 @@ try {
  * Fetches all components from styleguide and takes a screenshot of each.
  */
 const main = async () => {
-  const parser = new ArgumentParser()
-
-  parser.addArgument('--mode', {
-    choices: ['react', 'stack', 'kss'],
-    defaultValue: 'react'
-  })
-  parser.addArgument('--screenshot-dir', {
-    required: true,
-    dest: 'screenshotDir',
-    type: pathArgument
-  })
-  parser.addArgument('--styleguide-url', {
-    required: true,
-    dest: 'styleguideUrl',
-    type: urlArgument
-  })
-  parser.addArgument('--kss-dir', {
-    required: true,
-    dest: 'kssDir',
-    type: pathArgument
-  })
-  parser.addArgument('--viewport', {
-    defaultValue: builtinViewports.desktop
-  })
-  parser.addArgument('--no-empty-screenshot-dir', {
-    action: 'storeFalse',
-    defaultValue: true,
-    dest: 'emptyScreenshotDir'
-  })
-  parser.addArgument('--component')
-  parser.addArgument('--cache-file', {
-    defaultValue: '/tmp/cozy-ui-e2e-screenshots-cache.json',
-    dest: 'cacheFile'
-  })
-  parser.addArgument('--no-cache', {
-    dest: 'cacheFile',
-    action: 'storeFalse'
-  })
-
+  const parser = makeParser()
   const config = await readConfig()
   const args = parser.parseArgs()
